+++ dacite/src/lib.rs
@@ -40,6 +40,7 @@ pub mod khr_win32_surface;
 pub mod khr_xcb_surface;
 pub mod khr_xlib_surface;
 pub mod nv_dedicated_allocation;
+pub mod nv_external_memory_capabilities;
 
 pub use vks::android_wrapper;
 pub use vks::mir_wrapper;